+++ R/model_options.R
@@ -140,3 +140,16 @@ stitch_options_template <- function(options_template) {
         paste(options_template, collapse = ", ")
         , ']}', collapse = "")
 }
+
+#' Validate model options
+#'
+#' @param data The set of input data for the model run
+#' @param options Key value list of model options
+#'
+#' @return TRUE if valid else throws an error
+#' @keywords internal
+do_validate_model_options <- function(data, options) {
+  list(
+    valid = scalar(naomi::validate_model_options(data, options))
+  )
+}